Refact: Rename common component to TitleBox
Changed the name of the common component from MessageWidget to TitleBox for improved clarity and to avoid confusion with the actual widget component.
